+++ example/spec/redeployment/LambdaRedeploySpec.js
@@ -2,6 +2,8 @@
 
 const {
   buildAndStartWorkflow,
+  getLambdaAliases,
+  getLambdaVersions,
   waitForCompletedExecution,
   LambdaStep
 } = require('@cumulus/integration-tests');
@@ -11,51 +13,80 @@ const {
   redeploy
 } = require('../helpers/testUtils');
 
-const { updateConfigObject } = require('../helpers/configUtils');
-
 const config = loadConfig();
+const fs = require('fs-extra');
+
+
 const lambdaStep = new LambdaStep();
 
-describe('When a workflow', () => {
-  afterAll(async () => {
-    // Restore deployment following all following tests
-    const updateConfig = { handler: 'index.handler' };
-    const lambdaName = 'VersionUpTest';
-    const lambdaConfigFileName = './lambdas.yml';
-    updateConfigObject(lambdaConfigFileName, lambdaName, updateConfig);
+describe('When a workflow is running and a new version of a workflow lambda is deployed', () => {
+  let workflowExecutionArn;
+  let workflowStatus;
+  let testVersionOutput;
+
+  let startVersions;
+  let endVersions;
+  let startAliases;
+  let endAliases;
+  let startVersionNumbers;
+  let endVersionNumbers;
+  let startAliasVersionNumbers;
+  let endAliasVersionNumbers;
+
+  const originalFile = './lambdas/versionUpTest/original.js';
+  const updateFile = './lambdas/versionUpTest/update.js';
+  const targetFile = './lambdas/versionUpTest/index.js';
+
+  const lambdaName = `${config.stackName}-VersionUpTest`;
+
+  beforeAll(async () => {
+    //Redeploy 'new' copy of initial lambda
+    fs.copySync(originalFile, targetFile);
+    await fs.appendFile(targetFile, `//${new Date()}`);
     await redeploy(config);
+
+    startVersions = await getLambdaVersions(lambdaName);
+    startAliases = await getLambdaAliases(lambdaName);
+
+    fs.copySync(updateFile, targetFile);
+    await fs.appendFile(targetFile, `//${new Date()}`);
+    workflowExecutionArn = await buildAndStartWorkflow(
+      config.stackName,
+      config.bucket,
+      'TestLambdaVersionWorkflow'
+    );
+    await redeploy(config);
+    workflowStatus = await waitForCompletedExecution(workflowExecutionArn);
+    testVersionOutput = await lambdaStep.getStepOutput(
+      workflowExecutionArn,
+      lambdaName
+    );
+
+    endVersions = await getLambdaVersions(lambdaName);
+    endAliases = await getLambdaAliases(lambdaName);
+    endVersionNumbers = endVersions.map((x) => x.Version).filter((x) => (x !== '$LATEST'));
+    startVersionNumbers = startVersions.map((x) => x.Version).filter((x) => (x !== '$LATEST'));
+    endAliasVersionNumbers = endAliases.map((x) => x.FunctionVersion);
+    startAliasVersionNumbers = startAliases.map((x) => x.FunctionVersion);
   });
 
-  describe('is running and a new version of a workflow lambda is deployed', () => {
-    let workflowExecutionArn = null;
-    let workflowStatus = null;
-    let testVersionOutput = null;
+  it('the workflow executes successfully', () => {
+    expect(workflowStatus).toEqual('SUCCEEDED');
+  });
 
-    beforeAll(async () => {
-      const updateConfig = { handler: 'update.handler' };
-      const lambdaName = 'VersionUpTest';
-      const lambdaConfigFileName = './lambdas.yml';
+  it('uses the original software version', () => {
+    expect(testVersionOutput.payload).toEqual({ output: 'Current Version' });
+  });
 
-      workflowExecutionArn = await buildAndStartWorkflow(
-        config.stackName,
-        config.bucket,
-        'TestLambdaVersionWorkflow'
-      );
-      updateConfigObject(lambdaConfigFileName, lambdaName, updateConfig);
-      await redeploy(config);
-      workflowStatus = await waitForCompletedExecution(workflowExecutionArn);
-      testVersionOutput = await lambdaStep.getStepOutput(
-        workflowExecutionArn,
-        lambdaName
-      );
-    });
+  it('creates a new Lambda Version', () => {
+    expect(Math.max(...endVersionNumbers) - Math.max(...startVersionNumbers)).toEqual(1);
+  });
 
-    xit('the workflow executes successfully', () => {
-      expect(workflowStatus).toEqual('SUCCEEDED');
-    });
+  it('creates an updated Alias', () => {
+    expect(Math.max(...endAliasVersionNumbers) - Math.max(...startAliasVersionNumbers)).toEqual(1);
+  });
 
-    xit('uses the original software version', () => {
-      expect(testVersionOutput.payload).toEqual({ output: 'Current Version' });
-    });
+  it('creates aliases for all deployed versions', () => {
+    expect(endAliasVersionNumbers.sort()).toEqual(endVersionNumbers.sort());
   });
 });